requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
reportlab==4.1.0
pdfplumber==0.10.3
ollama==0.1.6
beautifulsoup4==4.12.3
requests==2.31.0 